Windows compilation, with review comments
